parser/tests/tree_construction/passing.dat

#data
<!DOCTYPE html><html><head></head><body><b>JAKE</b></body></html>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       "JAKE"

#data
Test
#errors
(1,0): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "Test"

#data
<p>One<p>Two
#errors
(1,3): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <p>
|       "One"
|     <p>
|       "Two"

#data
Line1<br>Line2<br>Line3<br>Line4
#errors
(1,0): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "Line1"
|     <br>
|     "Line2"
|     <br>
|     "Line3"
|     <br>
|     "Line4"

#data
<html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<head>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<body>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head></head>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head></head><body>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head></head><body></body>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head><body></body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head></body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head><body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<head></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
</head>
#errors
(1,7): expected-doctype-but-got-end-tag
#document
| <html>
|   <head>
|   <body>

#data
</body>
#errors
(1,7): expected-doctype-but-got-end-tag element.
#document
| <html>
|   <head>
|   <body>

#data
</html>
#errors
(1,7): expected-doctype-but-got-end-tag element.
#document
| <html>
|   <head>
|   <body>

#data
<b><table><td><i></table>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,14): unexpected-cell-in-table-body
(1,25): unexpected-cell-end-tag
(1,25): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <i>

#data
<b><table><td></b><i></table>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,14): unexpected-cell-in-table-body
(1,18): unexpected-end-tag
(1,29): unexpected-cell-end-tag
(1,30): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <i>
|       "X"

#data
<h1>Hello<h2>World
#errors
(1,4): expected-doctype-but-got-start-tag
(1,13): unexpected-start-tag
(1,18): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <h1>
|       "Hello"
|     <h2>
|       "World"

#data
<a><p>X<a>Y</a>Z</p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,10): unexpected-start-tag-implies-end-tag
(1,10): adoption-agency-1.3
(1,24): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <p>
|       <a>
|         "X"
|       <a>
|         "Y"
|       "Z"

#data
<b><button>foo</b>bar
#errors
(1,3): expected-doctype-but-got-start-tag
(1,18): adoption-agency-1.3
(1,21): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|     <button>
|       <b>
|         "foo"
|       "bar"

#data
<!DOCTYPE html><span><button>foo</span>bar
#errors
(1,39): unexpected-end-tag
(1,42): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <span>
|       <button>
|         "foobar"

#data
<p><b><div><marquee></p></b></div>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,11): unexpected-end-tag
(1,24): unexpected-end-tag
(1,28): unexpected-end-tag
(1,34): end-tag-too-early
(1,35): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|     <div>
|       <b>
|         <marquee>
|           <p>
|           "X"

#data
<script><div></script></div><title><p></title><p><p>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,28): unexpected-end-tag
#document
| <html>
|   <head>
|     <script>
|       "<div>"
|     <title>
|       "<p>"
|   <body>
|     <p>
|     <p>

#data
<!--><div>--<!-->
#errors
(1,5): incorrect-comment
(1,10): expected-doctype-but-got-start-tag
(1,17): incorrect-comment
(1,17): expected-closing-tag-but-got-eof
#new-errors
(1:5) abrupt-closing-of-empty-comment
(1:17) abrupt-closing-of-empty-comment
#document
| <!--  -->
| <html>
|   <head>
|   <body>
|     <div>
|       "--"
|       <!--  -->

#data
<p><hr></p>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,11): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <hr>
|     <p>

#data
<select><b><option><select><option></b></select>X
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): unexpected-start-tag-in-select
(1,27): unexpected-select-in-select
(1,39): unexpected-end-tag
(1,48): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|     <option>
|       "X"

#data
<a><table><td><a><table></table><a></tr><a></table><b>X</b>C<a>Y
#errors
(1,3): expected-doctype-but-got-start-tag
(1,14): unexpected-cell-in-table-body
(1,35): unexpected-start-tag-implies-end-tag
(1,40): unexpected-cell-end-tag
(1,43): unexpected-start-tag-implies-table-voodoo
(1,43): unexpected-start-tag-implies-end-tag
(1,43): unexpected-end-tag
(1,63): unexpected-start-tag-implies-end-tag
(1,64): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|     <a>
|       <b>
|         "X"
|       "C"
|     <a>
|       "Y"

#data
<a X>0<b>1<a Y>2
#errors
(1,5): expected-doctype-but-got-start-tag
(1,15): unexpected-start-tag-implies-end-tag
(1,15): adoption-agency-1.3
(1,16): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       x=""
|       "0"
|       <b>
|         "1"
|     <b>
|       <a>
|         y=""
|         "2"

#data
<!-----><font><div>hello<table>excite!<b>me!<th><i>please!</tr><!--X-->
#errors
(1,7): unexpected-dash-after-double-dash-in-comment
(1,14): expected-doctype-but-got-start-tag
(1,41): unexpected-start-tag-implies-table-voodoo
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): unexpected-cell-in-table-body
(1,63): unexpected-cell-end-tag
(1,71): eof-in-table
#document
| <!-- - -->
| <html>
|   <head>
|   <body>
|     <font>
|       <div>
|         "helloexcite!"
|         <b>
|           "me!"
|         <table>
|           <tbody>
|             <tr>
|               <th>
|                 <i>
|                   "please!"
|             <!-- X -->

#data
<!DOCTYPE html><li>hello<li>world<ul>how<li>do</ul>you</body><!--do-->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <li>
|       "hello"
|     <li>
|       "world"
|       <ul>
|         "how"
|         <li>
|           "do"
|       "you"
|   <!-- do -->

#data
<!DOCTYPE html>A<option>B<optgroup>C<select>D</option>E
#errors
(1,54): unexpected-end-tag-in-select
(1,55): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "A"
|     <option>
|       "B"
|     <optgroup>
|       "C"
|       <select>
|         "DE"

#data
<
#errors
(1,1): expected-tag-name
(1,1): expected-doctype-but-got-chars
#new-errors
(1:2) eof-before-tag-name
#document
| <html>
|   <head>
|   <body>
|     "<"

#data
<#
#errors
(1,1): expected-tag-name
(1,1): expected-doctype-but-got-chars
#new-errors
(1:2) invalid-first-character-of-tag-name
#document
| <html>
|   <head>
|   <body>
|     "<#"

#data
</
#errors
(1,2): expected-closing-tag-but-got-eof
(1,2): expected-doctype-but-got-chars
#new-errors
(1:3) eof-before-tag-name
#document
| <html>
|   <head>
|   <body>
|     "</"

#data
</#
#errors
(1,2): expected-closing-tag-but-got-char
(1,3): expected-doctype-but-got-eof
#new-errors
(1:3) invalid-first-character-of-tag-name
#document
| <!-- # -->
| <html>
|   <head>
|   <body>

#data
<?
#errors
(1,1): expected-tag-name-but-got-question-mark
(1,2): expected-doctype-but-got-eof
#new-errors
(1:2) unexpected-question-mark-instead-of-tag-name
#document
| <!-- ? -->
| <html>
|   <head>
|   <body>

#data
<?#
#errors
(1,1): expected-tag-name-but-got-question-mark
(1,3): expected-doctype-but-got-eof
#new-errors
(1:2) unexpected-question-mark-instead-of-tag-name
#document
| <!-- ?# -->
| <html>
|   <head>
|   <body>

#data
<!
#errors
(1,2): expected-dashes-or-doctype
(1,2): expected-doctype-but-got-eof
#new-errors
(1:3) incorrectly-opened-comment
#document
| <!--  -->
| <html>
|   <head>
|   <body>

#data
<!#
#errors
(1,2): expected-dashes-or-doctype
(1,3): expected-doctype-but-got-eof
#new-errors
(1:3) incorrectly-opened-comment
#document
| <!-- # -->
| <html>
|   <head>
|   <body>

#data
<?COMMENT?>
#errors
(1,1): expected-tag-name-but-got-question-mark
(1,11): expected-doctype-but-got-eof
#new-errors
(1:2) unexpected-question-mark-instead-of-tag-name
#document
| <!-- ?COMMENT? -->
| <html>
|   <head>
|   <body>

#data
<!COMMENT>
#errors
(1,2): expected-dashes-or-doctype
(1,10): expected-doctype-but-got-eof
#new-errors
(1:3) incorrectly-opened-comment
#document
| <!-- COMMENT -->
| <html>
|   <head>
|   <body>

#data
</ COMMENT >
#errors
(1,2): expected-closing-tag-but-got-char
(1,12): expected-doctype-but-got-eof
#new-errors
(1:3) invalid-first-character-of-tag-name
#document
| <!--  COMMENT  -->
| <html>
|   <head>
|   <body>

#data
<?COM--MENT?>
#errors
(1,1): expected-tag-name-but-got-question-mark
(1,13): expected-doctype-but-got-eof
#new-errors
(1:2) unexpected-question-mark-instead-of-tag-name
#document
| <!-- ?COM--MENT? -->
| <html>
|   <head>
|   <body>

#data
<!COM--MENT>
#errors
(1,2): expected-dashes-or-doctype
(1,12): expected-doctype-but-got-eof
#new-errors
(1:3) incorrectly-opened-comment
#document
| <!-- COM--MENT -->
| <html>
|   <head>
|   <body>

#data
</ COM--MENT >
#errors
(1,2): expected-closing-tag-but-got-char
(1,14): expected-doctype-but-got-eof
#new-errors
(1:3) invalid-first-character-of-tag-name
#document
| <!--  COM--MENT  -->
| <html>
|   <head>
|   <body>
